fix: fixed error states in sendCommand
During my testing I frequently got in an connection error state.
The first error state this commit fixes is when sending a command to a bulb while it is still connecting (e.g. startup or reconnect) it will emit an 'error' event on the socket (Error: Connection to device broken) but would not attempt to reconnect to the device
Secondly, when sending multiple commands the ceiling light commands did frequently take more than 3000 milliseconds to execute triggering the 'Error sending command' error. However, most of the time this command would still come through indicating the connection was not broken. Therefore I increased the timeout which triggers this error to 6000 seconds.